% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGeCodonPrecision.R
\name{ChoiceVector}
\alias{ChoiceVector}
\title{Choice vector of a grammar.}
\usage{
ChoiceVector(LHS)
}
\arguments{
\item{LHS}{Vector of Integers. 
The left-hand side \code{G$LHS} of a grammar object \code{G}.}
}
\value{
Vector of the number of choices for non-terminal symbols.
}
\description{
Choice vector of a grammar.
}
\examples{
NT<-sample(5, 50, replace=TRUE)
ChoiceVector(NT)
}
\seealso{
Other Utility: 
\code{\link{mLCMG}()}
}
\concept{Utility}
