% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGelF.R
\docType{data}
\name{lFxegaGeGene}
\alias{lFxegaGeGene}
\title{The local function list lFxegaGeGene.}
\format{
An object of class \code{list} of length 23.
}
\usage{
lFxegaGeGene
}
\description{
The local function list lFxegaGeGene.
}
\keyword{datasets}
