% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGeCodonPrecision.R
\name{mLCMGCodonPrecision}
\alias{mLCMGCodonPrecision}
\title{mLCMG precision of codon.}
\usage{
mLCMGCodonPrecision(LHS, ...)
}
\arguments{
\item{LHS}{Vector of Integers. The left-hand side of a grammar object \code{G}.}

\item{...}{Unused. Needed for the common abstract interface of 
precision functions.}
}
\value{
Integer. The precision of a codon whose upper bound is larger than the
        least common multiple of the prime factors of the 
        vector of the available rules 
        for each non-terminal of a grammar.
}
\description{
mLCMG precision of codon.
}
\examples{
NT<-sample(5, 50, replace=TRUE)
mLCMGCodonPrecision(NT)
}
\seealso{
Other Precision: 
\code{\link{CodonPrecisionWithThreshold}()},
\code{\link{MinCodonPrecision}()}
}
\concept{Precision}
